added support for NIP-62 Request to Vanish events #80
